+++ src/main/java/sopt/makers/authentication/adapter/in/web/controller/user/UserApiController.java
@@ -1,7 +1,7 @@
 package sopt.makers.authentication.adapter.in.web.controller.user;
 
-import static sopt.makers.authentication.support.constant.SystemConstant.API_KEY_HEADER;
-import static sopt.makers.authentication.support.constant.SystemConstant.SERVICE_NAME_HEADER;
+import static sopt.makers.authentication.common.constant.SystemConstant.API_KEY_HEADER;
+import static sopt.makers.authentication.common.constant.SystemConstant.SERVICE_NAME_HEADER;
 
 import sopt.makers.authentication.adapter.in.web.common.BaseResponse;
 import sopt.makers.authentication.adapter.in.web.common.code.UserSuccess;
