+++ src/main/java/sopt/makers/authentication/support/security/filter/ApiKeyAuthenticationFilter.java
@@ -1,8 +1,8 @@
 package sopt.makers.authentication.support.security.filter;
 
+import static sopt.makers.authentication.common.constant.SystemConstant.API_KEY_HEADER;
+import static sopt.makers.authentication.common.constant.SystemConstant.SERVICE_NAME_HEADER;
 import static sopt.makers.authentication.domain.auth.exception.AuthFailure.INVALID_API_KEY;
-import static sopt.makers.authentication.support.constant.SystemConstant.API_KEY_HEADER;
-import static sopt.makers.authentication.support.constant.SystemConstant.SERVICE_NAME_HEADER;
 
 import sopt.makers.authentication.domain.auth.exception.AuthException;
 import sopt.makers.authentication.support.security.authentication.ApiKeyAuthentication;
