+++ src/main/java/sopt/makers/authentication/support/security/filter/JwtAuthenticationFilter.java
@@ -1,10 +1,10 @@
 package sopt.makers.authentication.support.security.filter;
 
+import static sopt.makers.authentication.common.constant.SystemConstant.WHITELIST_WILDCARD;
 import static sopt.makers.authentication.domain.auth.exception.AuthFailure.MISSING_AUTHORIZATION_HEADER;
-import static sopt.makers.authentication.support.constant.SystemConstant.WHITELIST_WILDCARD;
 
+import sopt.makers.authentication.adapter.out.jwt.service.JwtAuthAccessTokenService;
 import sopt.makers.authentication.domain.auth.exception.AuthException;
-import sopt.makers.authentication.support.jwt.service.JwtAuthAccessTokenService;
 import sopt.makers.authentication.support.security.authentication.ApiKeyAuthentication;
 import sopt.makers.authentication.support.security.authentication.CustomAuthentication;
